Split up movableGroup and movableSingle
